Modularize Account logic
